+++ src/mopidy/backend.py
@@ -283,8 +283,6 @@ class PlaybackProvider:
 
         :param track: the track to play
         """
-        if track.uri is None:
-            return False
         uri = self.translate_uri(track.uri)
         if uri != track.uri:
             logger.debug("Backend translated URI from %s to %s", track.uri, uri)