+++ src/mopidy/core/_history.py
@@ -40,8 +40,6 @@ class HistoryController:
         if track.name is not None:
             name_parts.append(track.name)
         name = " - ".join(name_parts)
-        if track.uri is None:
-            return
         ref = Ref.track(uri=track.uri, name=name)
         self._history.insert(0, (timestamp, ref))