+++ src/mopidy/core/_library.py
@@ -245,10 +245,8 @@ class LibraryController:
                 if result is not None:
                     validation.check_instance(result, Mapping)
                     for uri, tracks in result.items():
-                        # TODO: Consider making Track.uri field mandatory, and
-                        # then remove this filtering of tracks without URIs.
                         validation.check_instances(tracks, Track)
-                        results[uri] = [track for track in tracks if track.uri]
+                        results[uri] = tracks
 
         return results