+++ tests/models/test_artist.py
@@ -4,11 +4,13 @@ import pydantic
 import pytest
 
 from mopidy.models import Artist
+from mopidy.types import Uri
+from tests.factories import ArtistFactory
 
 
 def test_uri():
     uri = "an_uri"
-    artist = Artist(uri=uri)
+    artist = ArtistFactory.build(uri=uri)
     assert artist.uri == uri
     with pytest.raises(pydantic.ValidationError):
         artist.uri = None
@@ -16,7 +18,7 @@ def test_uri():
 
 def test_name():
     name = "a name"
-    artist = Artist(name=name)
+    artist = ArtistFactory.build(name=name)
     assert artist.name == name
     with pytest.raises(pydantic.ValidationError):
         artist.name = None
@@ -24,7 +26,7 @@ def test_name():
 
 def test_musicbrainz_id():
     mb_id = "0383dadf-2a4e-4d10-a46a-e9e041da8eb3"
-    artist = Artist(musicbrainz_id=mb_id)
+    artist = ArtistFactory.build(musicbrainz_id=mb_id)
     assert artist.musicbrainz_id == UUID(mb_id)
     with pytest.raises(pydantic.ValidationError):
         artist.musicbrainz_id = None
@@ -32,26 +34,34 @@ def test_musicbrainz_id():
 
 def test_invalid_kwarg():
     with pytest.raises(pydantic.ValidationError):
-        Artist(foo="baz")
+        ArtistFactory.build(foo="baz")
 
 
 def test_invalid_kwarg_with_name_matching_method():
     with pytest.raises(pydantic.ValidationError):
-        Artist(replace="baz")
+        ArtistFactory.build(replace="baz")
 
     with pytest.raises(pydantic.ValidationError):
-        Artist(serialize="baz")
+        ArtistFactory.build(serialize="baz")
 
 
 def test_repr():
     assert (
-        repr(Artist(uri="uri", name="name"))
+        repr(
+            Artist(
+                uri=Uri("uri"),
+                name="name",
+            )
+        )
         == "Artist(uri='uri', name='name', sortname=None, musicbrainz_id=None)"
     )
 
 
 def test_serialize():
-    assert Artist(uri="uri", name="name").serialize() == {
+    assert Artist(
+        uri=Uri("uri"),
+        name="name",
+    ).serialize() == {
         "__model__": "Artist",
         "uri": "uri",
         "name": "name",
@@ -59,7 +69,10 @@ def test_serialize():
 
 
 def test_serialize_falsy_values():
-    assert Artist(uri="", name="").serialize() == {
+    assert Artist(
+        uri=Uri(""),
+        name="",
+    ).serialize() == {
         "__model__": "Artist",
         "uri": "",
         "name": "",