+++ tests/models/test_ref.py
@@ -1,20 +1,21 @@
 import pydantic
 import pytest
 
-from mopidy.models import Ref, RefType
+from mopidy.models import Ref
+from mopidy.types import Uri
 
 
 def test_uri():
     uri = "an_uri"
-    ref = Ref.track(uri=uri, name="Foo")
+    ref = Ref.track(uri=Uri(uri), name="Foo")
     assert ref.uri == uri
     with pytest.raises(pydantic.ValidationError):
-        ref.uri = None
+        ref.uri = Uri("")
 
 
 def test_name():
     name = "a name"
-    ref = Ref.track(uri="uri", name=name)
+    ref = Ref.track(uri=Uri("uri"), name=name)
     assert ref.name == name
     with pytest.raises(pydantic.ValidationError):
         ref.name = None
@@ -22,25 +23,23 @@ def test_name():
 
 # TODO: add these for the more of the models?
 def test_del_name():
-    ref = Ref.track(uri="foo", name="foo")
+    ref = Ref.track(uri=Uri("foo"), name="foo")
     with pytest.raises(pydantic.ValidationError):
         del ref.name
 
 
-def test_invalid_kwarg():
-    with pytest.raises(pydantic.ValidationError):
-        Ref(foo="baz")
-
-
 def test_repr_without_results():
     assert (
-        repr(Ref(uri="uri", name="foo", type=RefType.ARTIST))
+        repr(Ref.artist(uri=Uri("uri"), name="foo"))
         == "Ref(uri='uri', name='foo', type=ARTIST)"
     )
 
 
 def test_serialize_without_results():
-    assert Ref.track(uri="uri", name=None).serialize() == {
+    assert Ref.track(
+        uri=Uri("uri"),
+        name=None,
+    ).serialize() == {
         "__model__": "Ref",
         "type": "track",
         "uri": "uri",
@@ -56,35 +55,35 @@ def test_type_constants():
 
 
 def test_album_constructor():
-    ref = Ref.album(uri="foo", name="bar")
+    ref = Ref.album(uri=Uri("foo"), name="bar")
     assert ref.uri == "foo"
     assert ref.name == "bar"
     assert ref.type == Ref.ALBUM
 
 
 def test_artist_constructor():
-    ref = Ref.artist(uri="foo", name="bar")
+    ref = Ref.artist(uri=Uri("foo"), name="bar")
     assert ref.uri == "foo"
     assert ref.name == "bar"
     assert ref.type == Ref.ARTIST
 
 
 def test_directory_constructor():
-    ref = Ref.directory(uri="foo", name="bar")
+    ref = Ref.directory(uri=Uri("foo"), name="bar")
     assert ref.uri == "foo"
     assert ref.name == "bar"
     assert ref.type == Ref.DIRECTORY
 
 
 def test_playlist_constructor():
-    ref = Ref.playlist(uri="foo", name="bar")
+    ref = Ref.playlist(uri=Uri("foo"), name="bar")
     assert ref.uri == "foo"
     assert ref.name == "bar"
     assert ref.type == Ref.PLAYLIST
 
 
 def test_track_constructor():
-    ref = Ref.track(uri="foo", name="bar")
+    ref = Ref.track(uri=Uri("foo"), name="bar")
     assert ref.uri == "foo"
     assert ref.name == "bar"
     assert ref.type == Ref.TRACK