+++ docs/changelog.rst
@@ -116,6 +116,9 @@ Models
 
 Changes to the data models may affect any Mopidy extension or client.
 
+- The ``Track`` and ``Playlist`` models now requires the ``uri`` field to always
+  be set. (Fixes: :issue:`2190`, PR: :issue:`2229`)
+
 - The models are now based on Pydantic data classes, which means:
 
   - All models fields and the ``replace()`` method should work as before, so
@@ -160,6 +163,10 @@ Changes to the Audio API may affect a few Mopidy backend extensions.
 
 - Moved :class:`mopidy.audio.PlaybackState` to
   :class:`mopidy.types.PlaybackState`.
+
+- The :func:`mopidy.audio.tags.convert_tags_to_track` function now requires the
+  track ``uri`` as an argument, so that it can construct valid
+  :class:`~mopidy.models.Track` objects.
 
 - Removed APIs only used by Mopidy-Spotify's bespoke audio delivery mechanism,
   which has not been used since Spotify shut down their libspotify APIs in